requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.8.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-pubsub==2.23.0
google-cloud-secret-manager==2.20.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
watchdog==5.0.2
Werkzeug==3.0.4
